feat: manual edit types file
